migrate one file to typescript
