--- app/shared/inputs/TextField.js
@@ -1,47 +1,0 @@
-'use client'
-
-import { InputAdornment, TextField as MuiTextField } from '@mui/material'
-import { ErrorOutlineRounded } from '@mui/icons-material'
-
-const ADORNMENTS = {
-  error: <ErrorOutlineRounded className="text-red" />,
-}
-
-/**
- * @typedef {Object} TextFieldProps
- * @property {(keyof ADORNMENTS | JSX.Element)[]} endAdornments Shorthand prop for rendering adornments, pass array of icon elements or strings
- */
-
-/**
- * Wrapper around MuiTextField component
- * @param {MuiTextFieldProps & TextFieldProps}} props
- * @example
- * <TextField value="something" endAdornments={[
- *    <CheckRounded className='text-green' />,
- *    'error'
- *   ]}
- * />
- */
-
-export default function TextField({ endAdornments, ...restProps }) {
-  return (
-    <MuiTextField
-      variant="outlined"
-      {...restProps}
-      slotProps={{
-        ...restProps.slotProps,
-        input: {
-          sx: { fontFamily: 'var(--outfit-font)' },
-          endAdornment: endAdornments?.length > 0 && (
-            <InputAdornment position="end">
-              {endAdornments.map(
-                (adornment) => ADORNMENTS[adornment] ?? adornment,
-              )}
-            </InputAdornment>
-          ),
-          ...restProps.slotProps?.input,
-        },
-      }}
-    />
-  )
-}
